requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==1.7.3
Django==5.0.1
django-debug-toolbar==4.2.0
django-tailwind==2.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
Jinja2==3.1.3
jinja2-time==0.2.0
MarkupSafe==2.1.4
packaging==23.2
poyo==0.5.0
psycopg==3.1.17
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-slugify==8.0.2
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
urllib3==2.1.0
